requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
DateTime==4.3
defusedxml==0.6.0
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-OpenID==1.2.5
Flask-Session==0.3.1
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
python-dotenv==0.10.3
python-socketio==4.4.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.12
urllib3==1.25.7
vk==2.0.2
Werkzeug==0.16.0
zope.interface==4.7.1

